pyproject.toml
[project]
name = "dspp-trabalho-pratico-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
